chore: fix clippy lints across sdk + byo-relay
New stable-rust clippy version flagged latent warnings. Fixes:

sdk-core:
• webauthn tests use unwrap() — allow at module level (matches existing
  pattern in other test modules).
• webdav test mock manually implements impl Future to match the trait
  shape — allow clippy::manual_async_fn on the tests module.
• s3 start_upload_session: drop redundant `total_size > 0` check
  (total_size > MAX_MULTIPART_OBJECT_BYTES already handles the zero
  case correctly).

byo-relay:
• enrollment_admin: .map_err closure → function pointer.
• enrollment_admin test: assert_eq!(x, true) → assert!(x).
• rate_limit: drop redundant u64 casts on libc::statvfs fields; add
  comment noting the cast was for 32-bit targets we no longer ship.
• relay_auth::consume_and_verify: allow clippy::result_unit_err. The
  opaque Err(()) is deliberate — mismatch-kind leakage is the threat.
• sftp_relay::handle_text_command: allow clippy::too_many_arguments.

No behavior change. `cargo test` green on both workspaces.

+++ byo-relay/src/rate_limit.rs
@@ -843,8 +843,11 @@ pub fn disk_usage_percent(path: &std::path::Path) -> Option<u8> {
                 return None;
             }
             let st = st.assume_init();
-            let total = (st.f_blocks as u64).saturating_mul(st.f_frsize as u64);
-            let free = (st.f_bavail as u64).saturating_mul(st.f_frsize as u64);
+            // libc::statvfs fields are already u64 on our supported targets
+            // (x86_64-unknown-linux-gnu, aarch64-unknown-linux-gnu); casts
+            // were there to bridge 32-bit platforms that we no longer ship.
+            let total = st.f_blocks.saturating_mul(st.f_frsize);
+            let free = st.f_bavail.saturating_mul(st.f_frsize);
             if total == 0 {
                 return Some(0);
             }
